moving add components and setup flag to SimContext
