+++ src/vivarium/interface/interactive.py
@@ -5,7 +5,7 @@ from vivarium.framework.plugins import PluginManager
 from vivarium.framework.engine import SimulationContext
 from vivarium.config_tree import ConfigTree
 
-from .utilities import run_from_ipython, log_progress, raise_if_not_setup, InteractiveError
+from .utilities import run_from_ipython, log_progress, raise_if_not_setup
 
 from typing import Mapping, Union, List
 
@@ -15,13 +15,11 @@ class InteractiveContext(SimulationContext):
     def __init__(self, configuration, components, plugin_manager=None):
         super().__init__(configuration, components, plugin_manager)
         self._initial_population = None
-        self._setup = False
 
     def setup(self):
         super().setup()
         self._start_time = self.clock.time
         self.initialize_simulants()
-        self._setup = True
 
     def initialize_simulants(self):
         super().initialize_simulants()
@@ -117,11 +115,6 @@ class InteractiveContext(SimulationContext):
         self.component_manager._components.remove(old_component)
         new_component.setup(self.builder)
         self.component_manager.add_components([new_component])
-
-    def add_components(self, component_list):
-        if self._setup:
-            raise InteractiveError('Components cannot be added to a simulation once it is setup.')
-        self.component_manager.add_components(component_list)
 
 
 def initialize_simulation(components: List, input_config: Union[Mapping, ConfigTree]=None,
